Changed all registers to be int64_t

+++ include/vm.h
@@ -150,10 +150,10 @@ typedef struct Registers {
     int64_t *RB;
     int64_t *RC;
     int64_t *RD;
-    int32_t *RE;
-    int32_t *RF;
-    int32_t *RG;
-    int32_t *RH;
+    int64_t *RE;
+    int64_t *RF;
+    int64_t *RG;
+    int64_t *RH;
 
     //compare_flag can have three values:
     //-1 - The left operand is bigger.
